+++ apps/langchain_agent/telephony_app.py
@@ -2,7 +2,8 @@ import logging
 import os
 import typing
 from fastapi import FastAPI
-from vocode.streaming.models.events import Event, EventType, TranscriptCompleteEvent
+from vocode.streaming.models.events import Event, EventType
+from vocode.streaming.utils.transcript import TranscriptCompleteEvent
 from vocode.streaming.utils import events_manager
 
 from vocode.streaming.telephony.config_manager.redis_config_manager import (
@@ -28,7 +29,7 @@ class EventsManager(events_manager.EventsManager):
             transcript_complete_event = typing.cast(TranscriptCompleteEvent, event)
             add_transcript(
                 transcript_complete_event.conversation_id,
-                transcript_complete_event.transcript,
+                transcript_complete_event.transcript.to_string(),
             )
