+++ vocode/streaming/utils/transcript.py
@@ -2,7 +2,7 @@ import time
 from typing import List
 from pydantic import BaseModel, Field
 from enum import Enum
-from vocode.streaming.models.events import Sender, TranscriptEvent
+from vocode.streaming.models.events import Sender, Event, EventType
 
 from vocode.streaming.utils.events_manager import EventsManager
 
@@ -65,3 +65,18 @@ class Transcript(BaseModel):
             events_manager=events_manager,
             conversation_id=conversation_id,
         )
+
+
+class TranscriptEvent(Event, type=EventType.TRANSCRIPT):
+    text: str
+    sender: Sender
+    timestamp: float
+
+    def to_string(self, include_timestamp: bool = False) -> str:
+        if include_timestamp:
+            return f"{self.sender.name}: {self.text} ({self.timestamp})"
+        return f"{self.sender.name}: {self.text}"
+
+
+class TranscriptCompleteEvent(Event, type=EventType.TRANSCRIPT_COMPLETE):
+    transcript: Transcript
